As the name of the branch implies, this is a big ass overhaul:
In general, this is a transition from using mdutils to using mold as a low-level library for molecular dynamics operations.
The goal is to remove all dependencies of mdutils and just use mold, but this is an ongoing process.
Some parts are still not working under the new version yet, for example ramachandran plot, which will get some love and attention soon tm.
Changed font to droid sans mono-space for better readability in text editor.

Added implot as a library.
Added ImGuiColorTextEdit as a library for the TextEditor.

script editor:
The old property interface and language left much to be desired. The introduction of the scripting editor is taking the place of the property window.
Under the hood alot of things have happned.

timeline:
- Changed underlying visualization to use Implot

distribution:
- Changed underlying visualization to use Implot
- Missing: some way of filtering what to show here.

density volume:
- Now shown in its own window with its own controller

+ alot more details